+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- remove unnecessary uses of jQuery
+- configure eslint to disallow use of jQuery in addon
+
 ## [1.2.1]
 ### Changed
 - bump to ember-arcgis-portal-services 1.7.1 & torii-provider-arcgis 2.0